requirements.txt
Pillow==10.0.0
PyPDF2==3.0.1
genanki==0.13.0
pytesseract==0.3.10